backend/app/schemas/auth.py
from pydantic import BaseModel, Field

from app.models.user import UserRole


class LoginRequest(BaseModel):
    username: str = Field(..., description="用户名或手机号")
    password: str


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"


class UserPublic(BaseModel):
    id: int
    username: str
    phone: str
    role: UserRole

    model_config = {"from_attributes": True}
